Fix to CR OFF
When CR is OFF, asas should ask the airplane to follow the AP
instructions. Otherwise, asas will tell airplane to follow what ever
command it told the airplance to do before the conflict started. This
was usually the creation states. This was only a problem if AP wanted to
do something different exactly at the same time as a conflict is
detected.