Fix for short commands being sometimes skipped.
The scenario file parser would consider a line is empty (or too short to be a command) if its length was shorter than 12. An example of a valid command that would be skipped in this situation is 00:00:01>FF , which has a length of 11.

As the length of the timestamp needs to be at least 9 (including the > symbol), the parser should check whether the length of a line is smaller than 10.